requirements.txt
pyaudio
keyboard
groq
python-dotenv
aioreactive
expression
numpy